Allow win threshold (COMPLETE)
Allow the player to enter a win threshold as a command line option.
Replace check_win function with one which uses a single array for sums,
and checks whether square is the same as the previous square in
row/col/diagonal, rather than adding totals.
Add linear search function for checking wins.