Allow a win threshold to be specified as 2nd arg.
Allow the player to enter a win threshold as a command line option.
Replace check_win function with one which uses a single array for sums,
and checks whether square is the same as the previous square in
row/col/diagonal, rather than adding totals.
Add linear search function for checking wins.

Merge of:
commit fb07452e1976d2044282dae9ab6270f21b26bd18
commit 7dfad9f57e97d039a8b064a6a0131a033daeab54